fix: AI 인터뷰어 인터페이스, 면접 배경 컴포넌트 분리 및 dynamic import를 활용하여 페이지 사이즈 최소화
